Wolfram Mathematica Finger Kinematics
The Wolfram Mathematica files compute the tendon excursions for the fingers of the TU Hand given the angle of the DIP, PIP, and the flexion/extension and abduction/adduction angles of the MCP. There are two files, one for when the lateral bands pass over the mechanical stop dorsally, and the other when they pass on the volar side. Both call extraFunctions.m and Function_to_Determine_Points_of_Tangency.m in order to run. See Schultz and Rake "A four-tendon robotic finger with tendon transmission inspired by the human extensor mechanism" in Bioinspiration and Biomimetics, for more info. Please cite this paper if the code is used.