Kinematic Models of the Finger
Wolfram Mathematica files that compute the tendon excursions of the flexor, extensor, and radial and ulnar interossei tendons given the pose of the finger (DIP and PIP angles, and the flexion-extension and abduction-adduction angles of the MCP). extraFunctions.m and Function_to_Determine_Points_of_Tangency.m are called by the .nb files. One of these files computes the excursions when the lateral bands are routed dorsal to the mechanical stop, the other when the excursions are routed volar to the mechanical stop. For more information, see Schultz and Rake, "A four-tendon robotic nger with tendon transmission inspired by the human extensor mechanism" in Bioinspiration and Biomimetics. Please cite this paper if the code is used in your application.